follow refs to compute style
