+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
   "libs/compiler-yew",
   "libs/cli",
   "libs/rs_common",
-  "libs/parser_web",
   "libs/editor",
   "libs/project",
   "libs/evaluator",
