+++ libs/designer/jest.config.js
@@ -1,0 +1,10 @@
+const config = {
+  projects: [
+    {
+      displayName: "test",
+      testMatch: ["**/*.test.ts"],
+    },
+  ],
+};
+
+module.exports = config;
