+++ libs/designer/package.json
@@ -7,7 +7,7 @@
     "build": "yarn build:pc && yarn webpack",
     "build:watch": "yarn build:pc && concurrently \"yarn build:pc --watch\" \"webpack --watch\"",
     "build:pc": "../../target/debug/paperclip_cli build",
-    "test": "jest",
+    "test": "jest --runInBand --testTimeout=500 --bail --env=jsdom",
     "start": "nodemon --watch \"../../target/debug/paperclip_cli\" --exec \"../../target/debug/paperclip_cli designer --port=8000\""
   },
   "publishConfig": {
@@ -20,13 +20,13 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-typescript": "^7.18.6",
     "@improbable-eng/grpc-web-node-http-transport": "^0.15.0",
+    "@paperclip-ui/workspace": "workspace:^",
     "@types/jest": "^29.2.3",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "babel-jest": "^29.3.1",
     "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^4.1.0",
-    "execa": "^5.1.1",
     "fs-extra": "^10.1.0",
     "get-port": "^3.2.0",
     "html-webpack-plugin": "^5.5.0",
