+++ libs/designer/test/controls/mock-designer.ts
@@ -1,14 +1,11 @@
 import { createEditorMachine } from "../../src/machine";
-import * as path from "path";
 import { DEFAULT_STATE, EditorState } from "../../src/machine/state";
-import * as fsa from "fs-extra";
-import * as execa from "execa";
 import getPort from "get-port";
-import { EditorEvent, editorEvents } from "../../src/machine/events";
+import { EditorEvent } from "../../src/machine/events";
 import { EventEmitter } from "events";
 import { NodeHttpTransport } from "@improbable-eng/grpc-web-node-http-transport";
 import { Machine } from "@paperclip-ui/common";
-import { designerEngineEvents } from "@paperclip-ui/designer/src/machine/engine/designer/events";
+import { startWorkspace } from "@paperclip-ui/workspace/lib/test_utils";
 
 export type Designer = {
   onEvent(listener: (event: EditorEvent) => void): () => void;
@@ -21,37 +18,15 @@ export const startDesigner = async (
   initialState: Partial<EditorState> = {},
   namespace: string = "tmp-workspace"
 ): Promise<Designer> => {
-  files["paperclip.config.json"] = JSON.stringify(
-    {
-      srcDir: ".",
-      moduleDirs: [],
-      compilerOptions: [
-        {
-          emit: ["yew.rs:rs", "css", "react.js:js", "react.d.ts:d.ts"],
-          assetOutDir: "assets",
-        },
-      ],
-    },
-    null,
-    2
-  );
-  const tmpDirectory = `/private/tmp/pc-workspace/${namespace}`;
-  try {
-    fsa.rmdirSync(tmpDirectory);
-  } catch (e) {}
-  fsa.mkdirpSync(tmpDirectory);
+  const port = await getPort();
+  const workspace = await startWorkspace({
+    files,
+    namespace,
+    port,
+  });
 
-  const savedPaths = {};
   const em = new EventEmitter();
 
-  for (const relativePath in files) {
-    const filePath = path.join(tmpDirectory, relativePath);
-    fsa.mkdirpSync(path.dirname(filePath));
-    fsa.writeFileSync(filePath, files[relativePath]);
-    savedPaths[relativePath] = filePath;
-  }
-
-  const port = await getPort();
   const machine = createEditorMachine(
     {
       host: `localhost:${port}`,
@@ -72,15 +47,10 @@ export const startDesigner = async (
     history: {
       pathname: "/",
       query: {
-        file: savedPaths["entry.pc"],
+        file: workspace.localFilesPaths["entry.pc"],
       },
     },
   });
-
-  const ws = execa.command(
-    `${__dirname}/../../../../target/debug/paperclip_cli designer --port=${port}`,
-    { cwd: tmpDirectory, stdio: "ignore" }
-  );
 
   const onEvent = (listener: (event: EditorEvent) => void) => {
     em.on("event", listener);
@@ -90,7 +60,7 @@ export const startDesigner = async (
   };
 
   const dispose = () => {
-    ws.cancel();
+    workspace.dispose();
   };
 
   return { machine, onEvent, dispose };
