--- libs/parser_web/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "paperclip_parser_web"
-version = "0.1.1"
-authors = ["Craig Condon <craig.j.condon@gmail.com>"]
-edition = "2021"
-exclude = []
-
-[lib]
-name = "paperclip_parser_web"
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-paperclip_parser = { path = "../parser", features = [] }
-paperclip_proto = { path = "../proto" }
-serde = "1.0.147"
-serde-wasm-bindgen = "0.4.5"
-wasm-bindgen = "0.2.83"
