+++ libs/workspace/.gitignore
@@ -1,2 +1,2 @@
-*.js
+lib
 !dist/README.md
