+++ libs/workspace/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "cargo build",
+    "build": "cargo build; tsc",
     "build:watch": "concurrently \"cargo watch -x build\" \"nodemon --watch \"../designer/dist\" --exec 'touch src/lib.rs'\"",
     "test": "jest"
   },
@@ -14,10 +14,14 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@improbable-eng/grpc-web-node-http-transport": "^0.15.0",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^4.1.0",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.1.0",
+    "get-port": "^3.2.0",
     "html-webpack-plugin": "^5.5.0",
     "mini-css-extract-plugin": "^2.6.1",
     "nodemon": "^2.0.20",
